Calc aura shape in shape layer
