+++ client/src/game/layers/fow.ts
@@ -1,5 +1,5 @@
-import { SyncMode, InvalidationMode } from "@/core/comm/types";
-import { GlobalPoint, LocalPoint, Ray } from "@/game/geom";
+import { InvalidationMode, SyncMode } from "@/core/comm/types";
+import { GlobalPoint, Ray } from "@/game/geom";
 import { Layer } from "@/game/layers/layer";
 import { layerManager } from "@/game/layers/manager";
 import { Settings } from "@/game/settings";
@@ -12,7 +12,6 @@ import { getVisionSources } from "@/game/visibility/utils";
 import { visibilityStore } from "../visibility/store";
 import { TriangulationTarget } from "../visibility/te/pa";
 import { computeVisibility } from "../visibility/te/te";
-import { circleLineIntersection, xyEqual } from "../visibility/te/triag";
 
 export class FOWLayer extends Layer {
     isVisionLayer = true;
@@ -225,8 +224,8 @@ export class FOWLayer extends Layer {
                     this.vCtx.arc(lcenter.x, lcenter.y, g2lr(aura.value + aura.dim), 0, 2 * Math.PI);
                     this.vCtx.fill();
                     ctx.drawImage(this.virtualCanvas, 0, 0);
-                    aura.lastPath = this.updateAuraPath(polygon, auraCircle);
-                    shape.invalidate(true);
+                    // aura.lastPath = this.updateAuraPath(polygon, auraCircle);
+                    // shape.invalidate(true);
                 }
             }
 
@@ -258,69 +257,5 @@ export class FOWLayer extends Layer {
 
             ctx.globalCompositeOperation = originalOperation;
         }
-    }
-
-    updateAuraPath(visibilityPolygon: number[][], auraCircle: Circle): Path2D {
-        const path = new Path2D();
-        const lCenter = g2l(auraCircle.center());
-        const lRadius = g2lz(auraCircle.r);
-        let firstAngle: number | null = null;
-        let lastAngle: number | null = null;
-
-        const ixs: LocalPoint[][] = [];
-
-        // First find all polygon segments that are actually relevant
-        for (const [i, p] of visibilityPolygon.map(p => GlobalPoint.fromArray(p)).entries()) {
-            const np = GlobalPoint.fromArray(visibilityPolygon[(i + 1) % visibilityPolygon.length]);
-            const pLoc = g2l(p);
-            const npLoc = g2l(np);
-            const ix = circleLineIntersection(auraCircle.center(), auraCircle.r, p, np).map(x => g2l(x));
-            if (ix.length === 0) {
-                // segment lies completely outside circle
-                if (!auraCircle.contains(p)) continue;
-                // segment lies completely inside circle
-                else ix.push(pLoc, npLoc);
-            } else if (ix.length === 1) {
-                // segment is tangent to circle, segment can be ignored
-                if (xyEqual(ix[0].asArray(), pLoc.asArray()) || xyEqual(ix[0].asArray(), npLoc.asArray())) continue;
-                if (auraCircle.contains(p)) {
-                    ix.unshift(pLoc);
-                } else {
-                    ix.push(npLoc);
-                }
-            }
-            // Check some bad cases
-            if (ixs.length > 0) {
-                const lastIx = ixs[ixs.length - 1];
-                if (xyEqual(lastIx[0].asArray(), ix[1].asArray()) && xyEqual(lastIx[1].asArray(), ix[0].asArray()))
-                    continue;
-                if (xyEqual(lastIx[0].asArray(), ix[0].asArray()) && xyEqual(lastIx[1].asArray(), ix[1].asArray()))
-                    continue;
-            }
-            ixs.push(ix);
-        }
-
-        if (ixs.length < 1) {
-            path.arc(lCenter.x, lCenter.y, lRadius, 0, 2 * Math.PI);
-            return path;
-        }
-
-        // If enough interesting edges have been found, cut the circle up.
-        for (const ix of ixs) {
-            const circleHitAngle = Math.atan2(ix[0].y - lCenter.y, ix[0].x - lCenter.x);
-            if (lastAngle === null) {
-                path.moveTo(ix[0].x, ix[0].y);
-                firstAngle = Math.atan2(ix[0].y - lCenter.y, ix[0].x - lCenter.x);
-            } else if (lastAngle !== circleHitAngle) {
-                path.arc(lCenter.x, lCenter.y, lRadius, lastAngle, circleHitAngle);
-            }
-            lastAngle = Math.atan2(ix[1].y - lCenter.y, ix[1].x - lCenter.x);
-            path.lineTo(ix[1].x, ix[1].y);
-        }
-        if (firstAngle && lastAngle) {
-            if (firstAngle !== lastAngle) path.arc(lCenter.x, lCenter.y, lRadius, lastAngle, firstAngle);
-        } else path.arc(lCenter.x, lCenter.y, lRadius, 0, 2 * Math.PI);
-
-        return path;
     }
 }
