+++ client/src/game/shapes/utils.ts
@@ -1,3 +1,4 @@
+import { InvalidationMode, SyncMode } from "@/core/comm/types";
 import { uuidv4 } from "@/core/utils";
 import {
     ServerAsset,
@@ -10,7 +11,7 @@ import {
     ServerShape,
     ServerText,
 } from "@/game/comm/types/shapes";
-import { GlobalPoint, Vector } from "@/game/geom";
+import { GlobalPoint, LocalPoint, Vector } from "@/game/geom";
 import { layerManager } from "@/game/layers/manager";
 import { Asset } from "@/game/shapes/asset";
 import { Circle } from "@/game/shapes/circle";
@@ -21,8 +22,9 @@ import { Shape } from "@/game/shapes/shape";
 import { Text } from "@/game/shapes/text";
 import { EventBus } from "../event-bus";
 import { gameStore } from "../store";
+import { g2l, g2lz } from "../units";
+import { circleLineIntersection, xyEqual } from "../visibility/te/triag";
 import { Polygon } from "./polygon";
-import { SyncMode, InvalidationMode } from "@/core/comm/types";
 
 export function createShapeFromDict(shape: ServerShape): Shape | undefined {
     let sh: Shape;
@@ -160,4 +162,69 @@ export function deleteShapes(): void {
 export function cutShapes(): void {
     copyShapes();
     deleteShapes();
+}
+
+export function updateAuraPath(visibilityPolygon: number[][], center: GlobalPoint, aura: number): Path2D {
+    const auraCircle = new Circle(center, aura);
+    const path = new Path2D();
+    const lCenter = g2l(auraCircle.center());
+    const lRadius = g2lz(auraCircle.r);
+    let firstAngle: number | null = null;
+    let lastAngle: number | null = null;
+
+    const ixs: LocalPoint[][] = [];
+
+    // First find all polygon segments that are actually relevant
+    for (const [i, p] of visibilityPolygon.map(p => GlobalPoint.fromArray(p)).entries()) {
+        const np = GlobalPoint.fromArray(visibilityPolygon[(i + 1) % visibilityPolygon.length]);
+        const pLoc = g2l(p);
+        const npLoc = g2l(np);
+        const ix = circleLineIntersection(auraCircle.center(), auraCircle.r, p, np).map(x => g2l(x));
+        if (ix.length === 0) {
+            // segment lies completely outside circle
+            if (!auraCircle.contains(p)) continue;
+            // segment lies completely inside circle
+            else ix.push(pLoc, npLoc);
+        } else if (ix.length === 1) {
+            // segment is tangent to circle, segment can be ignored
+            if (xyEqual(ix[0].asArray(), pLoc.asArray()) || xyEqual(ix[0].asArray(), npLoc.asArray())) continue;
+            if (auraCircle.contains(p)) {
+                ix.unshift(pLoc);
+            } else {
+                ix.push(npLoc);
+            }
+        }
+        // Check some bad cases
+        if (ixs.length > 0) {
+            const lastIx = ixs[ixs.length - 1];
+            if (xyEqual(lastIx[0].asArray(), ix[1].asArray()) && xyEqual(lastIx[1].asArray(), ix[0].asArray()))
+                continue;
+            if (xyEqual(lastIx[0].asArray(), ix[0].asArray()) && xyEqual(lastIx[1].asArray(), ix[1].asArray()))
+                continue;
+        }
+        ixs.push(ix);
+    }
+
+    if (ixs.length < 1) {
+        path.arc(lCenter.x, lCenter.y, lRadius, 0, 2 * Math.PI);
+        return path;
+    }
+
+    // If enough interesting edges have been found, cut the circle up.
+    for (const ix of ixs) {
+        const circleHitAngle = Math.atan2(ix[0].y - lCenter.y, ix[0].x - lCenter.x);
+        if (lastAngle === null) {
+            path.moveTo(ix[0].x, ix[0].y);
+            firstAngle = Math.atan2(ix[0].y - lCenter.y, ix[0].x - lCenter.x);
+        } else if (lastAngle !== circleHitAngle) {
+            path.arc(lCenter.x, lCenter.y, lRadius, lastAngle, circleHitAngle);
+        }
+        lastAngle = Math.atan2(ix[1].y - lCenter.y, ix[1].x - lCenter.x);
+        path.lineTo(ix[1].x, ix[1].y);
+    }
+    if (firstAngle && lastAngle) {
+        if (firstAngle !== lastAngle) path.arc(lCenter.x, lCenter.y, lRadius, lastAngle, firstAngle);
+    } else path.arc(lCenter.x, lCenter.y, lRadius, 0, 2 * Math.PI);
+
+    return path;
 }
